Change UINT definition from fat_fs library from unsigned int to unsigned long, this will avoid compiler to create 16 bit counts and pointers on 8bit microcontrollers.

+++ SDK/lib/lib/fat_fs/inc/integer.h
@@ -12,7 +12,7 @@
 
 /* These types must be 16-bit, 32-bit or larger integer */
 typedef int				INT;
-typedef unsigned int	UINT;
+typedef unsigned long	UINT;
 
 /* These types must be 8-bit integer */
 typedef signed char		CHAR;
